+++ doc/news.rst
@@ -23,7 +23,8 @@ News / Changelog
 - Archive rendering no longer removes backslash escapes.  (Closes #490)
 - Add support for mistune 3.x, making dependency on mistune>=3.0.
 - Fix the rendering of contents of blockquote in plaintext rendering
-  mode. (Fixes #393)
+  mode. (See #393)
+- Fixes the rendering of patches in markdown rendering mode. (Fixes #393)
 
 
 .. _news-1.3.8: